Better Date Parsing Handling
I was getting a lot of crashes trying to parse timezone offsets with the fast ISO parser. The vast majority of dates returned from the Internet Archive are GMT dates so I stopped checking for time zone offsets and limited it to standard GMT-only formats. Other formats can use the slower, but more accurate `ISO8601DateFormatter`.