Add support for `collections_raw` field
This returns only the "real" collections for an item. The `collection` field returns an aggregate of the "real" collections and other "meta" collections like all of the people that have favorited the item. The favorites in particular can make the payload significantly larger for items that have a lot of favorites.